factor out refspecOverrides hacks into own file
